[SPACETIME] Moved old spacetime to deprecated build. New Spacetime is working. Hover over the window to move the mass over the XZ plane.

+++ source/spacetime/clspacetime.h
@@ -12,18 +12,18 @@
 #endif
 
 typedef struct _cl_spacetime_grid_t {
-	CL_TYPE(float)  m_mass[GRID_SIZE];
-	CL_TYPE(float)  m_pull[GRID_SIZE];
-	CL_TYPE(float4) m_position[GRID_SIZE];
-	CL_TYPE(float4) m_budge[GRID_SIZE];
+	CL_TYPE(double)  m_mass[GRID_SIZE];
+	CL_TYPE(double)  m_pull[GRID_SIZE];
+	CL_TYPE(double4) m_position[GRID_SIZE];
+	CL_TYPE(double4) m_budge[GRID_SIZE];
 } cl_spacetime_grid_t;
 
 typedef struct _cl_body_t {
-	CL_TYPE(float4) m_position;
-	CL_TYPE(float4) m_velocity;
-	CL_TYPE(float4) m_acceleration;
-	CL_TYPE(float)  m_mass;
-	CL_TYPE(float)  m_radius;
+	CL_TYPE(double4) m_position;
+	CL_TYPE(double4) m_velocity;
+	CL_TYPE(double4) m_acceleration;
+	CL_TYPE(double)  m_mass;
+	CL_TYPE(double)  m_radius;
 } cl_body_t;
 
 #endif
